+++ app/src/main/java/com/example/chemicalx/Fragment_Tasks/TaskCategoryAdapter.java
@@ -1,4 +1,4 @@
-package com.example.chemicalx.Fragment_Todolist;
+package com.example.chemicalx.Fragment_Tasks;
 
 import android.content.Context;
 import android.view.LayoutInflater;
