+++ app/src/main/java/com/example/chemicalx/Fragment_Tasks/crollerTest/OnCrollerChangeListener.java
@@ -1,4 +1,4 @@
-package com.example.chemicalx.Fragment_Todolist.crollerTest;
+package com.example.chemicalx.Fragment_Tasks.crollerTest;
 
 public interface OnCrollerChangeListener {
     void onProgressChanged(Croller croller, int progress);
