+++ app/src/main/java/com/example/chemicalx/Fragment_Tasks/crollerTest/utilities/Utils.java
@@ -1,4 +1,4 @@
-package com.example.chemicalx.Fragment_Todolist.crollerTest.utilities;
+package com.example.chemicalx.Fragment_Tasks.crollerTest.utilities;
 
 import android.content.Context;
 import android.content.res.Resources;
