+++ app/src/main/res/layout/croller_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.example.chemicalx.Fragment_Todolist.crollerTest.Croller
+<com.example.chemicalx.Fragment_Tasks.crollerTest.Croller
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent" />
